Complete the book borrow, return , approve function with new exception and modication others

+++ book-network-api-service/src/main/java/com/fmahadybd/book_network_api_service/feedback/Feedback.java
@@ -22,9 +22,8 @@ import com.fmahadybd.book_network_api_service.common.BaseEntity;
 public class Feedback extends BaseEntity {
 
     @Column
-    private Double note;
+    private Integer note;
     private String comment;
-
     @ManyToOne
     @JoinColumn(name = "book_id")
     private Book book;
